Rebased main and modularized enhanced question
